+++ package.json
@@ -27,6 +27,7 @@
     "express": "^4.18.2",
     "express-openid-connect": "^2.16.0",
     "express-validator": "^7.0.1",
+    "morgan": "^1.10.0",
     "sequelize": "^6.32.1",
     "sqlite3": "^5.1.6"
   }
